+++ NexmoClient/Internal/NexmoCore/Utility/NXMUtils.m
@@ -1,8 +1,7 @@
 //
 //  NXMUtils.m
-//  StitchObjC
+//  NexmoClient
 //
-//  Created by Doron Biaz on 7/10/18.
 //  Copyright © 2018 Vonage. All rights reserved.
 //
 
@@ -18,10 +17,28 @@
 static NSMutableDictionary<NSString*, NSNumber*>* statusMap = nil;
 static NSMutableDictionary<NSString*, NSNumber*>* typeMap = nil;
 
+static NSString * const NexmoDeviceUuidKey = @"NexmoClientDeviceUuid";
+
 + (NSDate *)dateFromISOString:(NSString *)isoString {
     NSDateFormatter *isoDateFomatter = [[NSDateFormatter alloc] init];
     isoDateFomatter.dateFormat = @"yyyy-MM-dd'T'HH:mm:ss.SSSZ";
     return [isoDateFomatter dateFromString:isoString];
 }
+
++ (NSString *)nexmoDeviceId {
+    NSString *deviceUuid = @"";
+    @synchronized([self class]) {
+        deviceUuid = [[NSUserDefaults standardUserDefaults] stringForKey:NexmoDeviceUuidKey];
+        if ([deviceUuid length] > 0) {
+            return deviceUuid;
+        }
+        
+        deviceUuid = [[NSUUID UUID] UUIDString];
+        [[NSUserDefaults standardUserDefaults] setObject:deviceUuid forKey:NexmoDeviceUuidKey];
+    }
+    
+    return deviceUuid;
+}
+
 
 @end