+++ NexmoTestApp/AppCode/Communications/CommunicationsManager.m
@@ -88,12 +88,12 @@
 }
 
 - (void)enablePushNotificationsWithDeviceToken:(nonnull NSData *)deviceToken
-                                     isPushKit:(BOOL)isPushKit
+                                     pushKit:(nonnull NSData *)pushKit
                                      isSandbox:(BOOL)isSandbox
                                     completion:(void(^_Nullable)(NSError * _Nullable error))completion {
-    [self.client enablePushNotificationsWithDeviceToken:deviceToken
-                                              isPushKit:isPushKit
-                                              isSandbox:isSandbox
+    [self.client enablePushNotificationsWithPushKitToken:deviceToken
+                                   userNotificationToken:pushKit
+                                               isSandbox:isSandbox
                                       completionHandler:completion];
 }